% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recompile_and_lookup.R
\name{recompile_and_lookup}
\alias{recompile_and_lookup}
\title{Recompile and lookup}
\usage{
recompile_and_lookup(path)
}
\arguments{
\item{path}{character with the path to the `Makefile`}
}
\value{
a vector of strings with output of `hfst-fst2strings`
}
\description{
Run `make` in the folder that recompiles `.hfst` from `.lexd` and run `hfst-fst2strings` afterwards.
}
\author{
George Moroz <agricolamz@gmail.com>
}
