% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_hfst.R
\name{run_tests}
\alias{run_tests}
\title{Test the results}
\usage{
run_tests(path)
}
\arguments{
\item{path}{character with the path to the `Makefile`}
}
\value{
a vector of strings with output of `hfst-fst2strings`
}
\description{
Use ***_test.csv for checking results of hfst.
}
\author{
George Moroz <agricolamz@gmail.com>
}
